+++ packages/core/src/components/Chips/index.ts
@@ -1,1 +1,2 @@
 export { default as Chips, type ChipsProps } from "./Chips";
+export * from "./Chips.types";