+++ packages/core/CHANGELOG.md
@@ -3,6 +3,47 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.5.3](https://github.com/mondaycom/vibe/compare/@vibe/core@4.5.2...@vibe/core@4.5.3) (2026-07-12)
+
+**Note:** Version bump only for package @vibe/core
+
+
+
+
+
+## [4.5.2](https://github.com/mondaycom/vibe/compare/@vibe/core@4.5.1...@vibe/core@4.5.2) (2026-07-06)
+
+
+### Bug Fixes
+
+* **Dropdown:** accessibility fixes for searchable single & multi select ([#3413](https://github.com/mondaycom/vibe/issues/3413)) ([4693f16](https://github.com/mondaycom/vibe/commit/4693f16794c3aa805751809ecb343beb7f58932c))
+
+
+
+
+
+## [4.5.1](https://github.com/mondaycom/vibe/compare/@vibe/core@4.5.0...@vibe/core@4.5.1) (2026-07-01)
+
+
+### Bug Fixes
+
+* **MultiStepIndicator:** fix invisible text on dark type pending hover ([#3412](https://github.com/mondaycom/vibe/issues/3412)) ([31acc6a](https://github.com/mondaycom/vibe/commit/31acc6a1d6cbd9a030ef2657d044f1ee2d89eaae))
+
+
+
+
+
+# [4.5.0](https://github.com/mondaycom/vibe/compare/@vibe/core@4.4.0...@vibe/core@4.5.0) (2026-06-30)
+
+
+### Features
+
+* **Chips:** add size prop with small variant ([#3409](https://github.com/mondaycom/vibe/issues/3409)) ([5ef809a](https://github.com/mondaycom/vibe/commit/5ef809a9d7b8cc49d34280df650f43cbf3d1146f))
+
+
+
+
+
 # [4.4.0](https://github.com/mondaycom/vibe/compare/@vibe/core@4.3.0...@vibe/core@4.4.0) (2026-06-15)