+++ packages/docs/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.1.3](https://github.com/mondaycom/vibe/compare/@vibe/docs@4.1.2...@vibe/docs@4.1.3) (2026-07-12)
+
+**Note:** Version bump only for package @vibe/docs
+
+
+
+
+
+## [4.1.2](https://github.com/mondaycom/vibe/compare/@vibe/docs@4.1.1...@vibe/docs@4.1.2) (2026-07-06)
+
+
+### Bug Fixes
+
+* **Dropdown:** accessibility fixes for searchable single & multi select ([#3413](https://github.com/mondaycom/vibe/issues/3413)) ([4693f16](https://github.com/mondaycom/vibe/commit/4693f16794c3aa805751809ecb343beb7f58932c))
+
+
+
+
+
+## [4.1.1](https://github.com/mondaycom/vibe/compare/@vibe/docs@4.1.0...@vibe/docs@4.1.1) (2026-07-01)
+
+**Note:** Version bump only for package @vibe/docs
+
+
+
+
+
+# [4.1.0](https://github.com/mondaycom/vibe/compare/@vibe/docs@4.0.11...@vibe/docs@4.1.0) (2026-06-30)
+
+
+### Features
+
+* **Chips:** add size prop with small variant ([#3409](https://github.com/mondaycom/vibe/issues/3409)) ([5ef809a](https://github.com/mondaycom/vibe/commit/5ef809a9d7b8cc49d34280df650f43cbf3d1146f))
+
+
+
+
+
 ## [4.0.11](https://github.com/mondaycom/vibe/compare/@vibe/docs@4.0.10...@vibe/docs@4.0.11) (2026-06-15)
 
 **Note:** Version bump only for package @vibe/docs